adding verison 2 support for geotag
