--- assets/kinvey.properties
@@ -1,2 +1,0 @@
-app.key=__KINVEY_APP_KEY__
-app.secret=__KINVEY_APP_SECRET__
