+++ project.properties
@@ -11,8 +11,6 @@
 #proguard.config=${sdk.dir}/tools/proguard/proguard-android.txt:proguard-project.txt
 
 # Project target.
-target=android-17
-
-android.library.reference.3=../google-play-services_lib
-android.library.reference.1=../../jakewharton/JakeWharton-ActionBarSherlock-e5c2d1c/library
-android.library.reference.2=../android-library/KinveyLib
+target=android-16
+android.library.reference.1=../../android-lib
+android.library.reference.2=../kitchensink-android/gen-external-apklibs/com.actionbarsherlock_actionbarsherlock_4.2.0
