+++ res/layout/activity_geo_tag.xml
@@ -1,12 +1,12 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical" >
+              android:layout_width="match_parent"
+              android:layout_height="match_parent"
+              android:orientation="vertical" >
 
     <com.google.android.gms.maps.MapView
-        android:id="@+id/geotag_map_main"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent" />
+            android:id="@+id/geotag_map_main"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent" />
 
 </LinearLayout>
