+++ res/menu/activity_geo_tag.xml
@@ -1,12 +1,12 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android" >
 
     <item
-        android:id="@+id/menu_item_refresh"
-        android:icon="@drawable/ic_menu_refresh"
-        android:showAsAction="ifRoom|withText"
-        android:title="@string/menu_refresh"/>
+            android:id="@+id/menu_item_refresh"
+            android:icon="@drawable/ic_menu_refresh"
+            android:showAsAction="ifRoom|withText"
+            android:title="@string/menu_refresh"/>
     <item
-        android:id="@+id/menu_item_legal"
-        android:title="@string/menu_legal"/>
+            android:id="@+id/menu_item_legal"
+            android:title="@string/menu_legal"/>
 
 </menu>
